Move TestImageDistGit tests to dedicated file
